Round whitelist money fields to int at DTO boundary
Introduce MoneyRounding helper (away-from-zero) and apply it at the
mapping/service boundary for prices, base amounts, estimated prices and
process paid/charged amounts. Eliminates the ±1 EGP drift between BE and
FE caused by decimal/double cast chains; DB columns, entity types,
internal calculations, request DTOs, VoltyksFees, MinimumFee and Wallet
all remain decimal/double.

+++ Voltyks.Core/DTOs/Charger/ChargerDetailsDto.cs
@@ -28,19 +28,13 @@ namespace Voltyks.Core.DTOs.Charger
         public CapacityDto Capacity { get; set; }
 
 
-        public decimal PricePerHour { get; set; }
+        public int PricePerHour { get; set; }
         public string AdapterAvailability { get; set; }
 
-        public double PriceEstimated { get; set; }
+        public int PriceEstimated { get; set; }
 
-        /// <summary>
-        /// Base charging cost before fees (PricePerHour × ChargingTimeInHours)
-        /// </summary>
-        public decimal BaseAmount { get; set; }
+        public int BaseAmount { get; set; }
 
-        /// <summary>
-        /// Voltyks platform fees
-        /// </summary>
         public decimal VoltyksFees { get; set; }
 
         /// <summary>
