+++ Voltyks.Core/DTOs/Process/MyActivityDto.cs
@@ -5,8 +5,8 @@ namespace Voltyks.Core.DTOs.Process
         public int Id { get; set; }
         public int ChargerRequestId { get; set; }
         public string? Status { get; set; }
-        public decimal? AmountCharged { get; set; }
-        public decimal? AmountPaid { get; set; }
+        public int? AmountCharged { get; set; }
+        public int? AmountPaid { get; set; }
         public DateTime? DateCreated { get; set; }
         public DateTime? DateCompleted { get; set; }
         public bool IsAsChargerOwner { get; set; }
